Use 1-Byte Hints
Changes the way hints are stored internally to use less memory.
This is a significant performance improvement: the time taken on the
hint generation benchmark is reduced by ~37% and the time taken on the
overall guess generation test is reduced by ~50%.

This change removes support for Wordle variants using word lengths other
than 5.